requirements.txt
asgiref==3.8.1
Django==3.2
django-bootstrap5==22.2
flake8==6.0.0
flake8-isort==6.0.0
isort==5.13.2
lxml==5.2.1
mccabe==0.7.0
pillow==10.3.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-docx==1.1.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
